+++ app/repositories/IRepository.ts
@@ -1,0 +1,6 @@
+import * as Config from "../config";
+
+export interface IRepository {
+    fullName: string;
+    getFreshData: () => Promise<Config.RepositoryDataDocument>;
+}
